backend/requirements.txt
polars>=1.5.0
duckdb>=1.0.0
pyyaml>=6.0.3
pandas>=2.2.0
openpyxl>=3.1.2
pyarrow>=15.0.0
python-calamine>=0.5.3
python-dotenv>=1.1.1
pyvis>=0.3.2
dbt-duckdb>=1.8.0
